requirements-8gb.txt
# Core ML
torch==2.0.1
torchaudio==2.0.2
openai-whisper==20231117
transformers==4.36.0
sentence-transformers==2.2.2
accelerate==0.25.0

# Quantization (optional)
bitsandbytes==0.41.0

# Audio processing
librosa==0.10.0
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
pyaudio==0.2.11
noisereduce==3.0.0
scipy==1.11.4

# Vector search
faiss-cpu==1.7.4

# Summarization
sumy==0.11.0
yake==0.4.8
keybert==0.8.0

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
redis==5.0.1
sqlalchemy==2.0.23

# Utilities
psutil==5.9.6
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1
pillow==10.1.0

# Document processing
PyMuPDF==1.23.8
python-pptx==0.6.21
PyPDF2==3.0.1

# Meeting bot (optional)
selenium==4.15.0
webdriver-manager==4.0.1

# Monitoring
memory-profiler==0.61.0